Major reorganization
All classic and LCLSTaoModel related code moved
Examples reorganized
New Developer section
