+++ examples/classic/data/epics_snapshot_2018-03-06T11:22:45.000000-08:00.json
