+++ examples/classic/data/epics_snapshot_2018-03-06T14:21:29.000000-08:00.json
